removed overhead of context api to prioritize performance slightly better readibility
